+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/projecte/Configuration.kt
@@ -8,19 +8,19 @@ object Configuration : IForgeConfigHandler {
     private var enableTurtleUpgradeConfig: ForgeConfigSpec.BooleanValue? = null
     private var enablePocketUpgradeConfig: ForgeConfigSpec.BooleanValue? = null
 
-    val enableNaturesCompassTurtleUpgrade: Boolean
+    val enableTransmutationTabletTurtleUpgrade: Boolean
         get() = enableTurtleUpgradeConfig?.get() ?: true
 
-    val enableNaturesCompassPocketUpgrade: Boolean
+    val enableTransmutationTabletPocketUpgrade: Boolean
         get() = enablePocketUpgradeConfig?.get() ?: true
 
     override val name: String
-        get() = "naturescompass"
+        get() = "projecte"
 
     override fun addToConfig(builder: ForgeConfigSpec.Builder) {
-        enableTurtleUpgradeConfig = builder.comment("Enables usage of natures compass as turtle upgrade")
-            .define("enableNaturesCompassTurtleUpgrade", true)
-        enablePocketUpgradeConfig = builder.comment("Enables usage of natures compass as pocket upgrade")
-            .define("enableNaturesCompassPocketUpgrade", true)
+        enableTurtleUpgradeConfig = builder.comment("Enables usage of transmutation tablet as turtle upgrade")
+            .define("enableTransmutationTabletTurtleUpgrade", true)
+        enablePocketUpgradeConfig = builder.comment("Enables usage of transmutation tablet as pocket upgrade")
+            .define("enableTransmutationTabletPocketUpgrade", true)
     }
 }